x/python: rename class names
